+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/vaccination/core/CertificatePersonIdentifier.kt
@@ -1,10 +1,11 @@
 package de.rki.coronawarnapp.vaccination.core
 
+import de.rki.coronawarnapp.covidcertificate.exception.InvalidHealthCertificateException.ErrorCode.VC_DOB_MISMATCH
+import de.rki.coronawarnapp.covidcertificate.exception.InvalidHealthCertificateException.ErrorCode.VC_NAME_MISMATCH
+import de.rki.coronawarnapp.covidcertificate.exception.InvalidVaccinationCertificateException
 import de.rki.coronawarnapp.covidcertificate.test.TestCertificateDccV1
 import de.rki.coronawarnapp.covidcertificate.test.TestCertificateQRCode
 import de.rki.coronawarnapp.util.HashExtensions.toSHA256
-import de.rki.coronawarnapp.vaccination.core.certificate.InvalidHealthCertificateException.ErrorCode
-import de.rki.coronawarnapp.vaccination.core.certificate.InvalidVaccinationCertificateException
 import de.rki.coronawarnapp.vaccination.core.certificate.VaccinationDGCV1
 import de.rki.coronawarnapp.vaccination.core.qrcode.VaccinationCertificateQRCode
 import org.joda.time.LocalDate
@@ -37,15 +38,15 @@ data class CertificatePersonIdentifier(
     fun requireMatch(other: CertificatePersonIdentifier) {
         if (lastNameStandardized != other.lastNameStandardized) {
             Timber.d("Family name does not match, got ${other.lastNameStandardized}, expected $lastNameStandardized")
-            throw InvalidVaccinationCertificateException(ErrorCode.VC_NAME_MISMATCH)
+            throw InvalidVaccinationCertificateException(VC_NAME_MISMATCH)
         }
         if (firstNameStandardized != other.firstNameStandardized) {
             Timber.d("Given name does not match, got ${other.firstNameStandardized}, expected $firstNameStandardized")
-            throw InvalidVaccinationCertificateException(ErrorCode.VC_NAME_MISMATCH)
+            throw InvalidVaccinationCertificateException(VC_NAME_MISMATCH)
         }
         if (dateOfBirth != other.dateOfBirth) {
             Timber.d("Date of birth does not match, got ${other.dateOfBirth}, expected $dateOfBirth")
-            throw InvalidVaccinationCertificateException(ErrorCode.VC_DOB_MISMATCH)
+            throw InvalidVaccinationCertificateException(VC_DOB_MISMATCH)
         }
     }
 }
