+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/vaccination/core/qrcode/VaccinationQRCodeValidator.kt
@@ -2,8 +2,8 @@ package de.rki.coronawarnapp.vaccination.core.qrcode
 
 import dagger.Reusable
 import de.rki.coronawarnapp.coronatest.qrcode.QrCodeExtractor
-import de.rki.coronawarnapp.vaccination.core.certificate.InvalidHealthCertificateException.ErrorCode.VC_PREFIX_INVALID
-import de.rki.coronawarnapp.vaccination.core.certificate.InvalidVaccinationCertificateException
+import de.rki.coronawarnapp.covidcertificate.exception.InvalidHealthCertificateException.ErrorCode.VC_PREFIX_INVALID
+import de.rki.coronawarnapp.covidcertificate.exception.InvalidVaccinationCertificateException
 import timber.log.Timber
 import javax.inject.Inject
